+++ upython.h
@@ -4,6 +4,7 @@
 #include <furi.h>
 
 #define TAG "uPython"
+#define CLI "py"
 
 typedef enum {
     ActionNone,
@@ -14,13 +15,16 @@ typedef enum {
 } Action;
 
 extern Action action;
+extern FuriString* file_path;
 
 Action mp_flipper_splash_screen();
 bool mp_flipper_select_python_file(FuriString* file_path);
 
-void mp_flipper_repl_register();
-void mp_flipper_repl_unregister();
+void mp_flipper_cli_register(void* args);
+void mp_flipper_cli_unregister(void* args);
 
-void mp_flipper_repl_execute(Cli* cli, FuriString* args, void* ctx);
+void mp_flipper_cli(Cli* cli, FuriString* args, void* ctx);
+
+void mp_flipper_repl_execute(Cli* cli);
 
 void mp_flipper_file_execute(FuriString* file);